Add optional dependency on pykeops; update imports and object locations to minimize imports using keops
