+++ pyproject.toml
@@ -51,10 +51,11 @@ lint = [
 ]
 test = ["pytest", "pytest-cov", "coverage", "jupyter", "ipython < 8.23"]
 graph = ["networkx"]
+varifold = ["pykeops"]
 autograd = ["autograd >= 1.3"]
 pytorch = ["torch >= 1.9.1"]
 backends = ["geomstats[autograd, pytorch]"]
-opt = ["geomstats[graph]"]
+opt = ["geomstats[graph, varifold]"]
 dev = ["geomstats[test, lint, doc]"]
 all = ["geomstats[dev, opt, backends]"]
